[CWS] move constant fetching to own package
